200403_Project3_A-B_Testing
Analyze A/B Test Results

A/B tests are very commonly performed by data analysts and data scientists. It is important that you get some practice working with the difficulties of these

Table of Contents
Introduction
Part I - Probability
Part II - A/B Test
Part III - Regression

For this project, you will be working to understand the results of an A/B test run by an e-commerce website. Your goal is to work through this notebook to help the company understand if they should implement the new page, keep the old page, or perhaps run the experiment longer to make their decision.